Create Sensitivity analysis controller and service
