+++ src/main/java/org/gridsuite/study/server/controller/sensitivityanalysis/SensitivityAnalysisParametersController.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.gridsuite.study.server.controller.sensitivityanalysis;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.gridsuite.study.server.StudyApi;
+import org.gridsuite.study.server.service.sensitivityanalysis.SensitivityAnalysisService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
+
+import static org.gridsuite.study.server.StudyConstants.HEADER_USER_ID;
+
+/**
+ * @author Bassel El Cheikh <bassel.el-cheikh_externe at rte-france.com>
+ */
+
+@RestController
+@RequestMapping(value = "/" + StudyApi.API_VERSION + "/studies/{studyUuid}/sensitivity-analysis")
+@Tag(name = "Study server - Sensitivity analysis parameters")
+public class SensitivityAnalysisParametersController {
+    private final SensitivityAnalysisService sensitivityAnalysisService;
+
+    public SensitivityAnalysisParametersController(SensitivityAnalysisService sensitivityAnalysisService) {
+        this.sensitivityAnalysisService = sensitivityAnalysisService;
+    }
+
+    @GetMapping(value = "/parameters")
+    @Operation(summary = "Get sensitivity analysis parameters on study")
+    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis parameters")})
+    public ResponseEntity<String> getSensitivityAnalysisParameters(
+            @PathVariable("studyUuid") UUID studyUuid) {
+        return ResponseEntity.ok().body(sensitivityAnalysisService.getSensitivityAnalysisParameters(studyUuid));
+    }
+
+    @PostMapping(value = "/parameters")
+    @Operation(summary = "set sensitivity analysis parameters on study, reset to default ones if empty body")
+    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis parameters are set"),
+        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
+    public ResponseEntity<Void> setSensitivityAnalysisParameters(
+            @PathVariable("studyUuid") UUID studyUuid,
+            @RequestBody(required = false) String sensitivityAnalysisParameters,
+            @RequestHeader(HEADER_USER_ID) String userId) {
+        return sensitivityAnalysisService.setSensitivityAnalysisParameters(studyUuid, sensitivityAnalysisParameters, userId) ? ResponseEntity.noContent().build() : ResponseEntity.ok().build();
+    }
+}
